+++ src/models/ElasticsearchIndex.ts
@@ -1,5 +1,5 @@
-import prettyBytes from 'pretty-bytes'
 import { EsIndex } from '../composables/components/indices/IndicesTable.ts'
+import { prettyPrintByteString } from '../helpers/bytes.ts'
 
 export default class ElasticsearchIndex {
   index: string
@@ -40,15 +40,5 @@ function parseIntValue (value: string) {
     return parseInt(value)
   } catch (_e) {
     return value
-  }
-}
-
-function prettyPrintByteString (value: string | number): string {
-  if (!value) return ''
-
-  if (typeof value === 'number' && !isNaN(value)) {
-    return prettyBytes(value)
-  } else {
-    return `${value} B`
   }
 }